--- Assets/Prefabs/Player/Scrips/Bullets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c7955c96313004993a51dee217e01540
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
